requirements.txt
flask
pyserial
flask-socketio
python-socketio
eventlet